backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pillow==11.0.0
email-validator==2.1.0
